Update DateTimePicker to version 2.5.18.
