+++ docs/HISTORY.txt
@@ -5,6 +5,8 @@ Changelog
 1.2.2 (unreleased)
 ------------------
 
+- Update DateTimePicker to version 2.5.18. [mathias.leimgruber]
+
 - Add Plone 5.1 support [njohner]
