+++ ftw/datepicker/profiles/default/jsregistry.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0"?>
 <object name="portal_javascripts" meta_type="JavaScripts Registry" autogroup="False">
 
-    <javascript id="++resource++datetimepicker/js/datetimepicker-2.4.5/jquery.datetimepicker.js"
+    <javascript id="++resource++datetimepicker/js/datetimepicker-2.5.18/jquery.datetimepicker.js"
                 compression="safe" enabled="True"
                 insert-after="++resource++plone.app.jquery.js" />
 
     <javascript id="++resource++datetimepicker/js/datetimepicker_widget.js"
                 compression="safe" enabled="True"
-                insert-after="++resource++datetimepicker/js/datetimepicker-2.4.5/jquery.datetimepicker.js" />
+                insert-after="++resource++datetimepicker/js/datetimepicker-2.5.18/jquery.datetimepicker.js" />
 
 </object>
