--- ftw/datepicker/resources/js/datetimepicker-2.4.5/.gitignore
@@ -1,2 +1,0 @@
-$ cat .gitignore
-parse.php
