--- ftw/datepicker/resources/js/datetimepicker-2.4.5/README.md
@@ -1,28 +1,0 @@
-datetimepicker
-==============
-[Documentation][doc]
-
-
-jQuery Plugin Date and Time Picker
-
-DateTimePicker
-
-![ScreenShot](https://raw.github.com/xdan/datetimepicker/master/screen/1.png)
-
-DatePicker
-
-![ScreenShot](https://raw.github.com/xdan/datetimepicker/master/screen/2.png)
-
-TimePicker
-
-![ScreenShot](https://raw.github.com/xdan/datetimepicker/master/screen/3.png)
-
-Options to highlight individual dates or periods
-
-![ScreenShot](https://raw.github.com/Mingpao/datetimepicker/master/screen/4.png)
-
-![ScreenShot](https://raw.github.com/Mingpao/datetimepicker/master/screen/5.png)
-
-![ScreenShot](https://raw.github.com/Mingpao/datetimepicker/master/screen/6.png)
-
-[doc]: http://xdsoft.net/jqplugins/datetimepicker/
