--- ftw/datepicker/resources/js/datetimepicker-2.4.5/bower.json
@@ -1,52 +1,0 @@
-{
-  "name": "datetimepicker",
-  "version": "2.4.5",
-  "main": [
-    "jquery.datetimepicker.js",
-    "jquery.datetimepicker.css"
-  ],
-  "ignore": [
-    "**/screen",
-    "**/datetimepicker.jquery.json",
-    "**/*.png",
-    "**/*.txt",
-    "**/*.md",
-    "**/*.html",
-    "**/*.tpl",
-    "**/jquery.js"
-  ],
-  "keywords": [
-    "calendar",
-    "date",
-    "time",
-    "form",
-    "datetime",
-    "datepicker",
-    "timepicker",
-    "datetimepicker",
-    "validation",
-    "ui",
-    "scroller",
-    "picker",
-    "i18n",
-    "input",
-    "jquery",
-    "touch"
-  ],
-  "dependencies": {
-    "jquery": ">= 1.7.2"
-  },
-  "authors": [
-    {
-      "name": "Chupurnov Valeriy",
-      "email": "chupurnov@gmail.com",
-      "homepage": "http://xdsoft.net/contacts.html"
-    }
-  ],
-  "license": "MIT",
-  "homepage": "http://xdsoft.net/jqplugins/datetimepicker/",
-  "repository": {
-    "type": "git",
-    "url": "git://github.com:xdan/datetimepicker.git"
-  }
-}
