--- ftw/datepicker/resources/js/datetimepicker-2.4.5/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "jquery-datetimepicker",
-  "version": "2.4.5",
-  "description": "jQuery Plugin DateTimePicker it is DatePicker and TimePicker in one",
-  "main": "jquery.datetimepicker.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/xdan/datetimepicker.git"
-  },
-  "keywords": [
-    "jquery-plugin",
-    "calendar",
-    "date",
-    "time",
-    "datetime",
-    "datepicker",
-    "timepicker"
-  ],
-  "author": "Chupurnov <chupurnov@gmail.com> (http://xdsoft.net/)",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/xdan/datetimepicker/issues"
-  },
-  "homepage": "https://github.com/xdan/datetimepicker"
-}
