+++ ftw/datepicker/resources/js/datetimepicker-2.5.18/MIT-LICENSE.txt
