+++ ftw/datepicker/resources/js/datetimepicker-2.5.18/jquery.datetimepicker.css
@@ -16,6 +16,9 @@
 	box-sizing: border-box;
 	display: none;
 }
+.xdsoft_datetimepicker.xdsoft_rtl {
+	padding: 8px 0 8px 8px;
+}
 
 .xdsoft_datetimepicker iframe {
 	position: absolute;
@@ -71,6 +74,11 @@
 	float: left;
 	margin-left: 8px;
 }
+.xdsoft_datetimepicker.xdsoft_rtl .xdsoft_datepicker {
+	float: right;
+	margin-right: 8px;
+	margin-left: 0;
+}
 
 .xdsoft_datetimepicker.xdsoft_showweeks .xdsoft_datepicker {
 	width: 256px;
@@ -83,13 +91,18 @@
 	margin-left: 8px;
 	margin-top: 0;
 }
+.xdsoft_datetimepicker.xdsoft_rtl .xdsoft_timepicker {
+	float: right;
+	margin-right: 8px;
+	margin-left: 0;
+}
 
 .xdsoft_datetimepicker .xdsoft_datepicker.active+.xdsoft_timepicker {
 	margin-top: 8px;
 	margin-bottom: 3px
 }
 
-.xdsoft_datetimepicker .xdsoft_mounthpicker {
+.xdsoft_datetimepicker .xdsoft_monthpicker {
 	position: relative;
 	text-align: center;
 }
@@ -155,6 +168,12 @@
 	display: block;
 	margin-left: 14px;
 	margin-top: 7px;
+}
+.xdsoft_datetimepicker.xdsoft_rtl .xdsoft_timepicker .xdsoft_prev,
+.xdsoft_datetimepicker.xdsoft_rtl .xdsoft_timepicker .xdsoft_next {
+	float: none;
+	margin-left: 0;
+	margin-right: 14px;
 }
 
 .xdsoft_datetimepicker .xdsoft_timepicker .xdsoft_prev {
@@ -389,6 +408,10 @@
 	top: 0;
 	bottom: 0;
 	cursor: pointer;
+}
+.xdsoft_datetimepicker.xdsoft_rtl .xdsoft_scrollbar {
+	left: 0;
+	right: auto;
 }
 .xdsoft_scroller_box {
 	position: relative;
