+++ ftw/datepicker/upgrades/20180315104957_update_to_date_time_picker_2_5_18/cssregistry.xml
@@ -3,6 +3,7 @@
 
     <stylesheet
         id="++resource++datetimepicker/js/datetimepicker-2.5.18/jquery.datetimepicker.css"
-        />
+        insert-after="++resource++datetimepicker/js/datetimepicker-2.4.5/jquery.datetimepicker.css" />
+    <stylesheet id="++resource++datetimepicker/js/datetimepicker-2.4.5/jquery.datetimepicker.css" remove="True" />
 
 </object>
