+++ ftw/datepicker/upgrades/20180315104957_update_to_date_time_picker_2_5_18/upgrade.py
@@ -1,0 +1,9 @@
+from ftw.upgrade import UpgradeStep
+
+
+class UpdateToDateTimePicker2518(UpgradeStep):
+    """Update to DateTimePicker 2.5.18.
+    """
+
+    def __call__(self):
+        self.install_upgrade_profile()
